Rename delete on blob endpoint. (#107)
pr: https://github.com/ABFSDriver/AbfsHadoop/pull/107

Enable rename and delete over Blob endpoint. The endpoint doesnot support rename API and not directory-delete.

There shall be no change until client layer, wherein, in the AbfsBlobClient#rename(), we will create an instance of BlobRenameHandler, and in AbfsBlobClient#delete(), we will create an instance of BlobDeleteHandler. These two new classes would own the orchestration logic for doing rename and delete, also they would own the HDFS contract prechecks. DFS endpoint performs the contract checks but not blob endpoint, hence, the handler would have to do prechecks and the postchecks.

How to directory operations?
In both of them, there is a need of listing that needs to be done by the client and then take actions on them. But, there is a problem with iterative listing and then taking action:

Server returns results in batch of a fixed count (here 5000), so if we wait to get all blob properties it would take, and then processing of the blobs would take time further.
Lets say the number of blobs are very huge, it would result in OOM on the client.
For the above point, the pr introduces a logic of producer and consumer. Where, the producer would keep producing the blob properties on a separate thread and enqueue them in a queue. The consumer would be owning an executorservice, it would keep polling the queue, and keep submitting to the executorService. The pr would take care of the consumer lag should not be more than a configuration.